Pull out a load of models into a common module
