+++ lib/mockzilla-common/build.gradle.kts
@@ -1,3 +1,4 @@
+
 import com.apadmi.mockzilla.JavaConfig
 import com.apadmi.mockzilla.debugVersionFile
 import com.apadmi.mockzilla.extractVersion
@@ -9,13 +10,12 @@ import com.apadmi.mockzilla.ProjectConfig
 import java.util.Date
 
 plugins {
+
     kotlin("multiplatform")
     kotlin("plugin.serialization")
-    id("com.android.library")
-    id("maven-publish")
     id("com.google.devtools.ksp")
-    id("com.codingfeline.buildkonfig")
     id("publication-convention")
+    id("com.android.library")
 }
 
 group = ProjectConfig.group
@@ -25,7 +25,7 @@ kotlin {
     android {
         publishAllLibraryVariants()
     }
-    
+
     val xcf = XCFramework()
     listOf(
         iosX64(),
@@ -52,25 +52,16 @@ kotlin {
             dependencies {
                 /* Kotlin */
                 implementation(libs.kotlinx.coroutines.core)
-                api(project(":mockzilla-common"))
-
-                /* Ktor */
-                api(libs.ktor.server.core)
-                implementation(libs.ktor.server.cio)
-                implementation(libs.ktor.server.content.negotiation)
-                implementation(libs.ktor.serialization.kotlinx.json)
                 implementation(libs.ktor.client.core)
-                implementation(libs.ktor.client.cio)
-                implementation(libs.ktor.server.rate.limit)
 
                 /* Serialization */
                 implementation(libs.kotlinx.serialization.json)
 
-                /* Logging */
-                implementation(libs.kermit)
-
                 /* Date Time */
                 implementation(libs.kotlinx.datetime)
+
+                /* Logging */
+                implementation(libs.kermit)
             }
         }
         val commonTest by getting {
@@ -126,34 +117,7 @@ android {
     }
 }
 
-buildkonfig {
-    packageName = "$group.mockzilla"
-
-    defaultConfigs {
-        buildConfigField(STRING, "VERSION_NAME", version.toString())
-    }
-}
-
-ksp {
-    arg("mockative.stubsUnitByDefault", "true")
-}
-
 tasks.dokkaHtml {
     outputDirectory.set(File(System.getProperty("docsOutputDirectory", "temp")))
 }
 
-dependencies {
-    configurations
-        .filter { it.name.startsWith("ksp") && it.name.contains("Test") }
-        .forEach {
-            add(it.name, "io.mockative:mockative-processor:1.2.3")
-        }
-}
-
-tasks.getByPath("publishToMavenLocal").dependsOn(
-    tasks.register("updateDebugMockzillaVersion" ) {
-        val newVersion = versionFile.readText().trim() + "-${Date().toInstant().epochSecond}"
-        version = newVersion
-        debugVersionFile.writeText(newVersion)
-    }
-)
