+++ lib/mockzilla-common/src/commonMain/kotlin/com/apadmi/mockzilla/lib/internal/models/GlobalOverridesDto.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
  * @property delayVariance
  */
 @Serializable
-internal data class GlobalOverridesDto(
+data class GlobalOverridesDto(
     val failProbability: Int? = null,
     val delayMean: Int? = null,
     val delayVariance: Int? = null
