+++ lib/mockzilla-common/src/commonMain/kotlin/com/apadmi/mockzilla/lib/models/EndpointConfiguration.kt
