+++ lib/mockzilla-common/src/commonTest/kotlin/com/apadmi/mockzilla/lib/HttpStatusCodeSerializerTests.kt
@@ -1,13 +1,12 @@
-package com.apadmi.mockzilla.lib.internal.utils
+package com.apadmi.mockzilla.lib
 
-import io.ktor.http.*
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
+import com.apadmi.mockzilla.lib.internal.utils.HttpStatusCodeSerializer
+import io.ktor.http.HttpStatusCode
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
+import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class HttpStatusCodeSerializerTests {
     @Test
