+++ lib/mockzilla/src/commonMain/kotlin/com/apadmi/mockzilla/lib/internal/di/DependencyInjector.kt
@@ -7,11 +7,11 @@ import com.apadmi.mockzilla.lib.internal.service.DelayAndFailureDecisionImpl
 import com.apadmi.mockzilla.lib.internal.service.LocalCacheServiceImpl
 import com.apadmi.mockzilla.lib.internal.service.MockServerMonitorImpl
 import com.apadmi.mockzilla.lib.internal.utils.FileIo
-import com.apadmi.mockzilla.lib.models.MetaData
 import com.apadmi.mockzilla.lib.models.MockzillaConfig
 import com.apadmi.mockzilla.lib.service.AuthHeaderProvider
 
 import co.touchlab.kermit.Logger
+import com.apadmi.mockzilla.lib.models.MetaData
 
 /**
  * @property logger
