+++ lib/mockzilla/src/commonMain/kotlin/com/apadmi/mockzilla/lib/service/MockzillaLogWriter.kt
@@ -1,19 +1,8 @@
 package com.apadmi.mockzilla.lib.service
 
-import com.apadmi.mockzilla.lib.models.MockzillaConfig
-import com.apadmi.mockzilla.lib.models.MockzillaConfig.LogLevel.Companion.toLogLevel
-
 import co.touchlab.kermit.LogWriter
 import co.touchlab.kermit.Severity
-
-interface MockzillaLogWriter {
-    fun log(
-        logLevel: MockzillaConfig.LogLevel,
-        message: String,
-        tag: String,
-        throwable: Throwable? = null
-    )
-}
+import com.apadmi.mockzilla.lib.models.MockzillaConfig
 
 @Suppress("EXTENSION_FUNCTION_WITH_CLASS")
 internal fun MockzillaLogWriter.toKermitLogWriter() = object : LogWriter() {
@@ -26,3 +15,6 @@ internal fun MockzillaLogWriter.toKermitLogWriter() = object : LogWriter() {
         this@toKermitLogWriter.log(logLevel = severity.toLogLevel(), message, tag, throwable)
     }
 }
+
+internal fun MockzillaConfig.LogLevel.toKermitSeverity() = Severity.valueOf(name)
+internal fun Severity.toLogLevel() = com.apadmi.mockzilla.lib.models.MockzillaConfig.LogLevel.valueOf(name)
