+++ lib/mockzilla/src/commonTest/kotlin/com/apadmi/mockzilla/lib/internal/controller/LocalMockControllerTests.kt
@@ -52,14 +52,17 @@ class LocalMockControllerTests {
         given(localCacheServiceMock).coroutine { getLocalCache("my-id") }.thenReturn(null)
 
         /* Run Test */
-        val response = sut.handleRequest(MockzillaHttpRequest(
+        val response = sut.handleRequest(
+            MockzillaHttpRequest(
             "http://example.com/local-mock/my-id",
             emptyMap(),
             method = HttpMethod.Get
-        ))
+        )
+        )
 
         /* Verify */
-        assertEquals(MockzillaHttpResponse(
+        assertEquals(
+            MockzillaHttpResponse(
             statusCode = HttpStatusCode.Created,
             headers = mapOf("test-header" to "test-value"),
             body = "my response body"
@@ -90,14 +93,17 @@ class LocalMockControllerTests {
         )
 
         /* Run Test */
-        val response = sut.handleRequest(MockzillaHttpRequest(
+        val response = sut.handleRequest(
+            MockzillaHttpRequest(
             "http://example.com/local-mock/my-id",
             emptyMap(),
             method = HttpMethod.Get
-        ))
+        )
+        )
 
         /* Verify */
-        assertEquals(MockzillaHttpResponse(
+        assertEquals(
+            MockzillaHttpResponse(
             statusCode = HttpStatusCode.InternalServerError
         ), response)
     }
