+++ lib/mockzilla/src/commonTest/kotlin/com/apadmi/mockzilla/lib/internal/utils/LogSeverityTests.kt
@@ -1,9 +1,10 @@
 package com.apadmi.mockzilla.lib.internal.utils
 
 import com.apadmi.mockzilla.lib.models.MockzillaConfig
-import com.apadmi.mockzilla.lib.models.MockzillaConfig.LogLevel.Companion.toLogLevel
 
 import co.touchlab.kermit.Severity
+import com.apadmi.mockzilla.lib.service.toKermitSeverity
+import com.apadmi.mockzilla.lib.service.toLogLevel
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
