+++ lib/mockzilla/src/commonTest/kotlin/com/apadmi/mockzilla/testutils/IntegrationTestUtils.kt
@@ -3,7 +3,6 @@ package com.apadmi.mockzilla.testutils
 import com.apadmi.mockzilla.lib.internal.service.LocalCacheService
 import com.apadmi.mockzilla.lib.internal.utils.FileIo
 import com.apadmi.mockzilla.lib.internal.utils.createFileIoforTesting
-import com.apadmi.mockzilla.lib.models.MetaData
 import com.apadmi.mockzilla.lib.models.MockzillaConfig
 import com.apadmi.mockzilla.lib.models.MockzillaRuntimeParams
 import com.apadmi.mockzilla.lib.prepareMockzilla
@@ -12,6 +11,7 @@ import com.apadmi.mockzilla.lib.stopMockzilla
 
 import co.touchlab.kermit.Logger
 import co.touchlab.kermit.StaticConfig
+import com.apadmi.mockzilla.lib.models.MetaData
 import io.ktor.client.network.sockets.*
 
 import kotlinx.coroutines.delay
